[atualizar conexão com pool]
